requirements.txt
opencv-python
Pillow
gpiozero
gphoto2
pycups
requests
pyqt5
qrcode
flask
